+++ crates/rgbm/src/dataset.rs
@@ -18,7 +18,7 @@ impl FeatureBinner {
         }
     }
 
-    pub fn array_to_bins(&self, array: &dyn Array) -> Result<Vec<u16>, ArrowError> {
+    pub fn array_to_bins(&self, array: &dyn Array) -> Result<Vec<u8>, ArrowError> {
         match self {
             Self::Numeric(b) => b.array_to_bins(array),
             Self::Categorical(b) => b.array_to_bins(array),
@@ -29,7 +29,7 @@ impl FeatureBinner {
 /// Training dataset: stores binned features, labels, and optional weights.
 pub struct Dataset {
     /// `binned_features[j][i]` is the bin index for row `i`, feature `j`.
-    pub binned_features: Vec<Vec<u16>>,
+    pub binned_features: Vec<Vec<u8>>,
     pub feature_binners: Vec<FeatureBinner>,
     pub feature_names: Vec<String>,
     pub labels: Float64Array,
